add bit map for alphabet used
alphabet bit map was created a while ago for a previous verilog. The letters look hokey because I literally created them from 8x8 grids on graphing paper and then tweaked all of them by hand after viewing what they looked like when displayed on a screen.

this alphabet does not work well when letters are directly on top of eachother. They look okay when there is a page of space in between them, but that greatly limits the usefulness because that basically means you can only display 32 characters on the OLED